fix: improve error message on key field type mismatch (#5335)
* fix: improve error message on key field type mismatch

fixes: https://github.com/confluentinc/ksql/issues/5333

The error message being returned to the user if the key field identified via `WITH(KEY)` syntax had a different SQL type to ROWKEY column was misleading as it talked about `WITH(KEY)` being deprecated and didn't mention `PRIMARY KEY`s on tables.  This is mainly due to the 'any key name' feature being pulled from the 0.9 release.  This change reverts the error message to the old version and adds in the PRIMARY KEY bit.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>