feat: drop WITH(KEY) syntax (#5363)
* feat: drop WITH(KEY) syntax

fixes: https://github.com/confluentinc/ksql/issues/3537

implements: See [KLIP-25](https://github.com/confluentinc/ksql/blob/master/design-proposals/klip-25-removal-of-with-key-syntax.md)

This change removes the `WITH(KEY)` syntax which previously allowed users to specify a value column that could act as an alias for the key column. This allowed a more user friendly name to be used for the key column, at the expense of requiring a copy of the key data in the value.

With the new 'any key name' feature, the key columns themselves can be given appropriate names, removing the need for this aliasing functionality.  See [KLIP-25](https://github.com/confluentinc/ksql/blob/master/design-proposals/klip-25-removal-of-with-key-syntax.md) for more details.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>