chore: ensure physical plan has correct key column name  (#5413)
* chore: ensure physical plan has correct key column name

Noticed that when a key column is aliased in the projection of a persistent query, the physical plan has the old name for the key column.

This change fixes that by including the names of the key columns in the `StreamSelect` / `TableSelect` physical plan step types.  It's sufficient to track just the column names as they key data can not be changed in a projection, only the column names.

Currently, only a single key column is possible. However, the select step pojos take a list of column names in prep for multiple key columns.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>