refactor: prep for getting ride of JoinKey
This change moves the validation done to ensure columns referenced in the projection for a query out of the `AstSanitizer` / `Analyzer` and into the logical model.

At this point, only the validation of SELECT expressions has moved. The rest will move in follow up PRs.

To achieve this, the `FinalProjectNode` needs to NOT take a logical schema as a parameter, but instead build it once the select items have been validated.  This means logical plan nodes are no longer immutable.

This is necessary work to enable the removal of JoinKey udf, which is on a tight deadline, so code may not be as clean as we'd normally like. It can be cleaned up after the deadline.