chore: switch synthetic join key to use ROWKEY
This change switching the naming of the synthetic key column created by some joins to use the name `ROWKEY`, or `ROWKEY_1`, `ROWKEY_2` etc if the is a name clash.

The change also contains tidy up and refactoring. Specifically, realizing that the validation `Analyzer` does is shared with pull queries, so it doesn't make sense to move this validation.  Instead, given the synthetic key column is a short to medium term problem only, i.e. it will go away with time, switched the design so that all validation returns to `Analyzer`, except is now won't through on an unknown column is it _might_ be a synthetic key column.  Validation of potential synthetic key columns are handled by the logic model. This works because pull queries don't (yet) have this problem.

 - There was duplicate logic between `AstSantizer` and `Analayzer` around reporting ambiguous and unknown columns.
    - Changed so that, for now, `AstSanitzer` reports ambiguous columns and `Analyzer` reports unknown columns. Not perfect, but better than both attempting to do both.
 - It puts back the validation of Selects in `Analyser` and enhanced it and `ColumnReferenceValidator` to know when a synthetic key column is possible and to not throw on one. False positives are possible, but are caught by the logical model.
 - Simplified the logical model to only worry about validating synthetic key columns in projections.
 - Split `RepartitionNode` into classes specific for its two use-cases.