refactor: remove duplicate code & unneeded validation from logical plan
This change removes some duplication by adding an abstract `SingleSourcePlanNode` class to handle the common case of nodes that only have a single source node.

It also simplifies `PreJoinProjectNode` by moving the creation of the schema and select expressions into the class itself, removing the need to validate they match,
and simplifies the `FinalProjectNode`  by also removing unnecessary validation that the schema matches the selects, which is not needed as the class builds these.

The `PreJoinProjectNode` could also be simplified more as its not actually mutating the data, only the schema of the data, but this is left to another PR.