chore: do not through if stream without key column has data in key
If a stream with no key column is created over a topic:

```sql
CREATE STREAM S (NAME STRING) WITH (...);
```

and if the underlying topic's records has data in the key, then any use of the stream was failing with the error "Data should be null for a VoidDeserializer.".

This failure was because we were using the Stream's inbuilt `Serdes.Void()` serde to handle the key of streams with no key columns. However, the deserializer was throwing this error. The fix is to use our own deserializer which ignores any data in the topic record.

This also highlighted a hole in the test framework: if a JSON test case does not define the `key` field of input and output records it is defaulted to an empty string. This is wrong, it should default to `null`, which it now does.  A few existing tests, including historical ones, needed fixing up after this change.

To allow test cases to provide data for the key of streams that have no column, the test framework will serialize any `key` within an input record, where the associated stream does not have a key column, as a `STRING`.