test: enhance test framework to use clickable links in errors
Like others, I spend a lot of time opening and searching through QTT & RQTT test files for the right test when a test fails, sometimes getting the wrong one and wasting time.  Enter this change... This change adds clickable links to the output of th tests.

For example, if the `entries sorted` test fails in the `array.json` file then you'll get a link like:

```
file:///Users/andy/dev/github.com/confluentinc/ksql_2/ksqldb-functional-tests/src/test/resources/query-validation-tests/array.json:7
```

Clicking this link in IntelliJ will take you to the right line in the `array.json` file.

If a historic test fails, you'll get a link to the historic tests directory. Click the link in IntelliJ will take you to the correct test's directory.

All of this is achieved by the addition of a `TestLocation` interface, two implementations for either a location with a line number or without, and a `TestFileContext` which tracks all of this and allows a test name to be converted to a location in a file.

Time saver!