fix: improve join predictability by increasing max task idle
improvement for: https://github.com/confluentinc/ksql/issues/5537

This change increases the Stream's `max.task.idle.ms` from its default `0` to `500`ms, which should provide a better out-of-the-box join experience for anyone starting a join with a large backlog of messages, i.e. a backlog large enough to mean the broker may only return records for one side of the join. See #5537 for more info.