feat: include row count in SHOW TOPICS EXTENDED
While investigating issues I've always thought it would be useful if the output of `SHOW TOPICS EXTENDED` included the number of records within the topic.

With this change, it now does... ish!

The record count is calculated as the difference between the earliest and latest offsets. For non-compacted topics this shows the number of records in the topic.
For compacted topics this shows the maximum number of records that can be in the topic, though compaction may have reduced this number.

### old output:

```
ksql> SHOW TOPICS EXTENDED;

 Kafka Topic                    | Partitions | Partition Replicas | Consumers | ConsumerGroups
----------------------------------------------------------------------------------------------
 default_ksql_processing_log    | 1          | 1                  | 0         | 0
 pageviews                      | 1          | 1                  | 0         | 0
 users                          | 1          | 1                  | 0         | 0
----------------------------------------------------------------------------------------------
```

### new output:

```
ksql> SHOW TOPICS EXTENDED;

 Kafka Topic                    | Partitions | Replicas | Max Row Count | Consumers | ConsumerGroups
----------------------------------------------------------------------------------------------------
 default_ksql_processing_log    | 1          | 1        | 0             | 0         | 0
 pageviews                      | 1          | 1        | 2345          | 0         | 0
 users                          | 1          | 1        | 25            | 0         | 0
-----------------------------------------------------------------------------------------------------
```