fix: support GROUP BY with no source columns used
fixes:

Adds support for a `GROUP BY` query where there are no source columns used in the query, e.g.

```sql
CREATE TABLE OUTPUT as SELECT 1 as k, count(1) AS ID FROM INPUT group by 1;
```

This was previously failing as it ran into an issue deserializing the query plan due to `nonAggColumns` not being present in the plan.