refactor: clean up TableRowsEntity (#5651)
* refactor: clean up TableRowsEntity

fixes: https://github.com/confluentinc/ksql/issues/3865

Quick refactor: the old TableRowsEntity derived from KsqlEntity and had code in the CLI to handle printing it.  This was because pull queries were originally handled by our POST admin commands. This is no longer the case. So this change simplifies `TableRowsEntity` by removing the super class, moving it out of the ksqldb-rest-client-model module, renaming to `TableRows`, and cleaning up the code.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>