fix: allow joins in windowed aggregations
fixes: #5898

This change fixes a regression introduced in v5.5.0 that meant any windowed aggregation with a join would fail with an `IllegalArgumentException`.

This change fixes the regression, how ever follow on work is required to allow access to the window bounds columns `WINDOWSTART` and `WINDOWEND` in such queries.  Access to these columns was not possible in v5.4, i.e. this is not a regression. The follow on work will be tracked under #5931.