fix: avoid losing cause of processing errors
This change removes code that was dropping the first 'cause' of a processing logger error.

I *think* the code previously did this because the `errorMsg` was taken from the first 'cause' of the passed in `Exception`. However, the code now accepts a custom `errorMsg`, which may not match the first 'cause' of the supplied exception.

For example:

```java
// Old code:
} catch (final Exception e) {
   // msg.errorMsg is first line of of 'e.cause',
   // msg.cause is all other lines of 'e.cause'
   final ErrorMessage msg = RecordProcessingError.recordProcessingError(e, row);
   ...
}

// New code:
} catch (final Exception e) {
   // msg.errorMsg is "Error doing something"
   // msg.cause drops first line of 'e.cause' ... which is dropping useful information!
   final ErrorMessage msg = RecordProcessingError.recordProcessingError("Error doing something", e, row);
   ...
}
```

There may still be some uses where the first line of the cause duplicates the error msg. However, better to have duplicate data than to drop important information.