fix: NPE when udf metrics enabled (#5960)
* fix: NPE when udf metrics enabled

fixes: https://github.com/confluentinc/ksql/issues/5890

Fixes an NPE thrown when the second instance of a UDAF is initialized. The NPE was being thrown due to a bug in the code that failed to initialise the sensor fields in the `UdafAggregateFunction` class if the sensor already existed.

The commit also refactors the code used to add function invocation metrics such that all three function types: UDF, UDTF and UDAF, use the same function to register metrics.

Unit tests added to ensure this single code path doesn't throw an NPE or return null on the subsequent invocation.

Also added `ksql.udf.collect.metrics` metric to the server config file, so users are more likely to be able to find it and try it out.

Moved UDTF functions out of the `ksql-udf` group and into `ksql-udtf` group. Having them in the former was another bug.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>