test: fix flaky test due to static state
The `TlsTest` tended to fail locally when running other tests in parallel. Investigation showed this is due to shared state in `ServerKeyStore`: the path to the key store.  The test needed to write a bad cert and then assert requests failed, but if another test ran it would override the bad cert with a good one.