refactor: better fix for protobuf not supporting unwrapped values
fixes: https://github.com/confluentinc/ksql/issues/5994

Better fix for the above issue. Previous fix didn't fit right.

This fix introduces a `SerdeFeature` enum and an associated `supportedFeatures()` method on `Format`. This allows formats to define the features it supports.  This will work well when we come to 'User defined formats'.

If the user supplies `WRAP_SINGLE_VALUES` in the `WITH` clause, this will map to either `SerdeOption.WRAP_SINGLE_VALUES` or `SerdeOption.UNWRAP_SINGLE_VALUES`.  These two options map to `SerdeFeature.WRAP_SINGLES` and `SerdeFeature.UNWRAP_SINGLES`, respectively.

If the user doesn't supply `WRAP_SINGLE_VALUES` then no option is set and the format is free to use its default format. Though this requires a follow up change to allow the default to be 'unwrapped'.

The default value for the config that provides the default for single value wrapping has been removed. This is to allow formats to define what they do by default, rather than having it enforced by the system. For example, JSON will wrap by default, where as KAFKA will be unwrapped by default.