feat: Support for IF NOT EXISTS on CREATE CONNECTOR (#6036)
`CREATE CONNECTOR` now accepts an optional `IF NOT EXISTS` clause. 

When excluded, the statement WILL fail if a connector with the same name already exists.

When included, the statement will NOT fail because a connector with the same name already exists. Note: ksqlDB does not e ensure the existing connector has the same config as declared in the statement.

```sql
-- Only create connector 'BOB' if a connector with the same name does not already exist:
CREATE SOURCE CONNECTOR IF NOT EXISTS BOB WITH (...);
```