test: avoid the need to 'correct' KsqlStatementExceptions in QTT
Negative test cases should if it is not the last statement in the test case that causes the expected exception to be thrown.  This is so that there aren't test cases were the author thinks it is the last statement that fails, but its actually an earlier one.

To achieve this, the code was testing the contents of the `KsqlStatementException` to assert it matched the last statement in the test case.  However, because the engine reformts code it the test executor needed to mess around correcting the contents of the statement exception.

This commit simplifies things by just asserting the last statement failed, rather than relying on `KqlStatementException` contents.