fix: fail on non-string MAP keys (#6182)
* fix: fail on non-string MAP keys

fixes: https://github.com/confluentinc/ksql/issues/6181

Fixes ksqlDB so that statements that create maps with non-string keys fail.

Crux of the issue was in `SqlTypeParser`, which ignored the key type.

Additionally, there were many places in the code that checked for non-STRING keys, but they always were STRING keys by this point.  These checks have generally been removed in prep for supporting non-STRING keys.  It is the format that should decide if it can handle non-STRING keys or not, hence the code now fails these statements because the key/value format rejects them as unsupported. This opens the way to PROTOBUF supporting integer keys in the future: only the Protobuf Serde could should need to change. (https://github.com/confluentinc/ksql/issues/6177).

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>