fix: register correct unwrapped schema (#6188)
* fix: register correct unwrapped schema

This commit fixes several issues and refactors a lot of the serde code around wrapping and unwrapping single values.

The main issues being fixed are:
  1. allow each format to define if it supported wrapping and/or unwrapping. (Not possible with current design)
  2. pass the correct wrapping / unwrapping flags are passed to key vs value formats when creating serde. (bug in code passes same SerdeOptions to key and value).
  3. register the correct wrapped / unwrapped schema with the SR. (bug in existing code meant registered format is always wrapped).

At the same time, the way wrapping / unwrapping was handled in the code wasn't great. Formats like `JSON` needed to be able to handle both wrapped and unwrapped schemas and values, depending on whether the user _explicitly_ set wrapping or unwrapping, vs the default behaviour of the format. This commit refactors the code such that the format will always be passed the a consistent schema and the set of serde features the format should use when creating the serde. This simplifies things and paves the way to user-define-serde.


Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>