refactor: push Connect schema down (#6200)
* refactor: push ConnectSchema down

Following on from the fixes and refactors done in #6188, this commit pushes down the use of the Connect schema type to lower levels of the code. Higher levels of the code now deal with `LogicalSchema`, `PersistentSchema` or just a `List<SimpleColumn>`.

This moves us closer to removing the Connect schema from the code base, except in the Serde code that deals with connect formats.

`LogicalSchema` and `PersistentSchema` no longer know about the Connect schema type. Calls to retrieve the Connect schema from these types have been replaced with a util function that can convert a list of columns into a Struct Connect schema.  As more code moves away from the Connect schema these util function calls will slowly be removed.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>