refactor: introduce SessionConfig
There's a fair amount of code that passes around `KsqlConfig`, holding the system config, paired with a `Map<String, Object>`, containing the overrides the user supplied with their request.  Recipients of these parameters can apply the overrides to the config to get a new `KsqlConfig` instance with the overrides applied by calling `cloneWithPropertyOverwrite`.  However, I've fixed a few bugs previously where this was missing, or simplified it where was done multiple times.

To address these, I've bitten the bullet and spent and hour introducing a `SessionConfig` class, which holds the system config and overrides.  This type can be passed where ever the two configs were previously being passed.

The `getConfig` method takes a boolean control whether overrides are applied or not.  I chose this over two separate methods, e.g. `getSystemConfig` and `getConfigWithOverrides`, as my hope is that the boolean will force anyone writing new code to think about whether they want the config with or without overrides.