refactor: format interface
In prep for user-defined-formats and just to tidy up the interface and fix some bugs...

## Refactor

Refactored the `Format` interface so that the functions used during schema inference, i.e. `toParsedSchema` and `toColumns`, are moved onto a new interface `SchemaTranslator`. The `SchemaTranslator` is obtained from the `Format` by calling `getSchemaTranslator`, passing in the format properties.

This removes the strange requirement of having to pass a `FormatInfo`, which contains the name of the format, to the `toParsedSchema` method.

And ensures consistent properties are used for both `toParsedSchema` and `toColumns`. (Though at the moment properties are only needed in `toParsedSchema`, that may not be the case once user-definfed-formats are supported.

## Fixes / Enhancements
Fixed a bug / ux issue, where a C* statement that was using schema inference would not report any issues if the supplied `VALUE_FORMAT` did not match the type of the schema in the schema registry. For example, if `VALUE_FORMAT` was `AVRO`, but the schema returned by the Schema Registry was `JsonSchema`, then the statement would succeed... and would later generate lots of deserialisation errors if used. The statement will now fail.