fix: JSON format should correct scale of decimals when deserializing
It is the responsibility of the format to ensure the data returned matches the required schema. This includes the scale of decimals. The JSON format was not correctly setting the scale of decimals when deserializing. For example, give:

```sql
CREATE STREAM S (ID INT KEY, PRICE DECIMAL(10,2) WITH (kafka_topic='S', formats='JSON');
```

The above creates a stream with a single value column `PRICE` which should have a scale of `2`, i.e. two decimal places.

If the data in tbe Kafka record's value was to have the incorrect, e.g.

```json
{
   "price": 12
}
```

or

```json
{
   "price": 12.1
}
```

Then the deserializer was returning the decimal as provided, i.e. `12` or `12.1`. However, this is incorrect as the schema of the column states is has a scale of two. So all values for the column should have the scale set to two, i.e. the above examples should deserialize to `12.00` and `12.10`. With this change they no do.