chore: ensure CLI works with other key formats
fixes: https://github.com/confluentinc/ksql/issues/6218

Updates a few places in the rest api, cli and the new Java client to specify the key format as well as the value format.

Unfortunately, the value format was called `format`.  This change introduces a `keyFormat` and `valueFormat` to hold the two formats, but both the CLI and new Java client are backwards compatible with the old JSON schema.

Also in this commit, streams also indicate if there key is windowed.  This means the key is serialized using the windowed serializer. It is possible to have a stream over a topic with a windowed key. Now the CLI and Java client expose this info.

Main change was to `show tables` and `show streams` response.  Previously, these returned

```
ksql> show streams;

 Stream Name | Kafka Topic | Format
---------------------------------------
 PAGEVIEWS   | pageviews   | DELIMITED
---------------------------------------

ksql> show tables;

 Table Name | Kafka Topic | Format | Windowed
----------------------------------------------
 USERS      | users       | JSON   | false
----------------------------------------------
```

And now return:

```
ksql> show streams;

 Stream Name | Kafka Topic | Key Format | Value Format | Windowed
-----------------------------------------------------------------
 PAGEVIEWS   | pageviews   | KAFKA      | DELIMITED    | false
-----------------------------------------------------------------

ksql> show tables;

 Table Name | Kafka Topic | Key Format | Value Format | Windowed
----------------------------------------------------------------
 USERS      | users       | KAFKA      | JSON          | false
----------------------------------------------------------------
```