refactor: change QTT to store serialized data in mock Kafka
Switch QTT from storing deserialized keys and values in the mock Kafka service to serialized.

This switcharoo means the test cases don't need to know the serialization formats before the data has been run through the Kafka Streams topology.  A later change will reply on this fact to leverage the information ksqlDB gleans when messages flow through the topology to fix an issue with the `TopicInfoCache`.