Topic info cache (#6366)
The `TopicInfoCache` is used by QTT to retrieve info about the topics being used by the test case. Specifically, its used to get at the schema, key and value formats of the topic, and potentially use this information to build appropriate serde.

The implementation wasn't correctly handling internal topics. When building information about internal topics it was using the key and value formats of the sink topic. This is often incorrect.  For example, consider a windowed group by where the source needs to be repartitioned, i.e. the `GROUP BY` is not on the key. In such a situation the repartition topic doesn't have a windowed key, which the sink topic does.

To get the correct key format, the `TopicInfoCache` now gets the information from the query itself. Mainly the `QuerySchemas`, which holds the logical schema plus the key and value formats used when creating the serde passed to Kafka Streams.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>