test: topology change: state store dropped
The upstream AK commit https://github.com/apache/kafka/pull/9156 seems to have had an unexpected side affect, which is causing QTT to fail because the topology has changed.

With the new streams snapshot the topology is no longer materializing the `KTABLE-SOURCE-000000000X` nodes.  This looks to be a good thing, as the next node is also materialized, making the first materialization redundant.