test: fix SQL tester tests (#6494)
Fixes some test failures due to upstream changes.

With the upstream fix https://github.com/apache/kafka/pull/9156 in, the streams topology is no longer emitting tombstones for rows that were never in the table to start with when an input doesn't pass a filter.

This change was causing some SQL Tester tests to fail.  This change fixes those tests.

The downside of the upstream change is that query upgrades can now miss sending tombstones for any rows that where in the table _before_ the upgrade, but which don't belong in the table _after_ the upgrade, e.g. because a filter has changed.  See https://github.com/confluentinc/ksql/issues/6493 for more details.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>