refactor: remove unused validation code
`KsqlStruct` is not yet used. However, the code calls `validateValue` when setting the value of a field. This check is not free and is on the critical path.  ksqlDB is strongly typed, so the only way the type would be wrong would be if a function or deserializer where returning incorrect data, i.e. a bug.  We should not pay the production runtime cost of checking _every_ field we set during _every_ record we process.

Rather than validate here, we should add a `-debug` option to ksqlDB that validates the returned data from deserializers and functions against the expected schema, as suggested here: https://github.com/confluentinc/ksql/issues/3624#issuecomment-729532558