chore: switch to schema-less KsqlStruct (#6659)
* refactor: remove unused validation code

`KsqlStruct` is not yet used. However, the code calls `validateValue` when setting the value of a field. This check is not free and is on the critical path.  ksqlDB is strongly typed, so the only way the type would be wrong would be if a function or deserializer where returning incorrect data, i.e. a bug.  We should not pay the production runtime cost of checking _every_ field we set during _every_ record we process.

Rather than validate here, we should add a `-debug` option to ksqlDB that validates the returned data from deserializers and functions against the expected schema, as suggested here: https://github.com/confluentinc/ksql/issues/3624#issuecomment-729532558

* chore: switch to schema-less KsqlStruct

No other sql instance types, (e.g. BigDecimal, String, Double, Map, List, etc), require the schema during instantiation. Only the Connect `Struct` type and the ksqlDB specific `KsqlStruct` type, which is to be `Struct`'s replacement, require the schema during Construction.  This can make working with structs more tricky than other types.

This change removes the requirement to provide the schema when instantiating `KsqlStruct`. The aim is to make this type easier to use.

The `Field` type has also been moved into the `SqlStruct` class. The code base has many `Field` types. `Field` is specifically a struct field. Making this an inner class of `SqlStruct` helps avoid confusion as to which field type we are talking about.

* chore: fix merge issue

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>