bound the size of the shared buffer cache used by queries
bound the total size of the shared buffer cache used by queries. The
bound is controlled by 2 configs, one for controlling the limit for
persistent queries, the other for transient queries. Today, this limit
is implemented by accounting the full limit for each query against the
total bound. In the future we may implement by using a shared cache.