fix: fix how the buffer limit check evaluates streams config
The buffer limit check first creates a dummy streams config to get the current
max bytes buffering value. To do this it creates a map of the streams props and
creates the config (with dummy values for all the required fields). However
sometimes those fields are already provided and ImmutableMap doesn't like
overwrites in the builder. So, use a plain old HashMap