fix: fix the cache max bytes buffering check (#7181)
* fix: fix the cache max bytes buffering check

this patch fixes our validation of cache.max.bytes.buffering. We validate based
on the set of live queries in the engine context. However, we weren't popullating
this set for sandboxes. This is fixed in this change by introducing a sandbox type
for transient query metadata, and filling in the live queries set with sandboxed
query metadatas

* fix checkstyle

* fix test