refactor(executor): turn RemoteDataAugmenter into an Executor
There is really no reason for this class to handle the transform.
Instea, let each call site ow modify the results returned from
RemoteHostExecutor in situ as they please.

It's fine to return the first KsqlEntity always because we run the
executor on singular statements, so KsqlEntityList
always has length 1