feat(executor): Allow SourceDescription returning executors to return cluster wide stream stats (#7252)
* feat: add random-beans libraries

These give a simple implementation of object mother pattern. Again
this should help us simplify testing a bit

* refactor(executor): turn RemoteDataAugmenter into an Executor

There is really no reason for this class to handle the transform.
Instea, let each call site ow modify the results returned from
RemoteHostExecutor in situ as they please.

It's fine to return the first KsqlEntity always because we run the
executor on singular statements, so KsqlEntityList
always has length 1

* feat(executor): add an executor for fetching remote source descriptions

* feat(executor): encapsulate cluster wide query metrics

ClusterQueryStats encapsulates the cluster wide
query stats for the entire cluster. It comes with one
constructor for assembling the stats divided by host

* fix(test): get to the bottom of the test flakes

A few problems fixed in this commit:
1. some tests were missing the correct mockito test runner - this
could cause trouble if some of the code called out to by these uses
mockito mocks
2. rules need to be public and annotated, otherwise they cause races
3. remove static declaration of side effects - also caused races

* feat(executor): collate cluster wide source stats on all ListSources

This means SourceDescription queries will be able to give you a per-node
breakdown of stats available on the new fields:
- clusterStatistics
- clusterErrorStats